requirements.txt
yfinance
pandas
python-dotenv
schedule
requests
certifi
